:bug: Fix small bugs.

+++ build.js
@@ -1,18 +1,20 @@
 'use strict';
 
-const {mkdirSync, writeFileSync} = require('fs');
 const crx3 = require('crx3');
 const {generateKeyPairSync} = require('crypto');
+const {mkdirSync, writeFileSync, existsSync} = require('fs');
 
-const {publicKey, privateKey} = generateKeyPairSync('rsa', {
-  modulusLength: 4096,
-  publicKeyEncoding: {type: 'spki', format: 'pem'},
-  privateKeyEncoding: {type: 'pkcs8', format: 'pem'},
-});
+if (!existsSync('dist')) {
+  const {publicKey, privateKey} = generateKeyPairSync('rsa', {
+    modulusLength: 4096,
+    publicKeyEncoding: {type: 'spki', format: 'pem'},
+    privateKeyEncoding: {type: 'pkcs8', format: 'pem'},
+  });
 
-mkdirSync('dist');
-writeFileSync('dist/id_rsa', privateKey);
-writeFileSync('dist/id_rsa.pub', publicKey);
+  mkdirSync('dist');
+  writeFileSync('dist/id_rsa', privateKey);
+  writeFileSync('dist/id_rsa.pub', publicKey);
+}
 
 crx3(['src/manifest.json'], {
   keyPath: 'dist/id_rsa',
